+++ tests/utils/mod.rs
@@ -1,0 +1,5 @@
+// SPDX-FileCopyrightText: 2022 Shun Sakai
+//
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+
+pub mod command;
